Add meta descriptions (#21)
* chore: override the `host` configuration in Svelte Kit

We need have the correct base URL configured inside the application, so
that it can be output in the markup. To do that, we set the expected
host name in the `.env` files, and then grab the value from the
environmetn variable in `svelte.config.js`.

* feat: introduce MetaDescriptors component

MetaDescriptors component outputs meta tags for the description, and the
basic open graph protocol properties.

It automatically suffixes the title, and bases the `og:url` value on
what comes from `$app/stores.page`.

* feat: use MetaDescriptors component on every page

Uses the MetaDescriptors component on every page, so that every page has
the basic meta tags.