fix(modgrid): restore scroll position of modgrid after navigation
Previously the scroll position of the modgrid would be reset when a user navigated back
to the grid page. This is due to the scrolling setup, where `div#page` is scrolled rather
than the complete page.

That could probably be properly fixed in the future, but this works for now.